+++ jax/experimental/array_api/__init__.py
@@ -36,12 +36,12 @@ and implements most of the API listed in the standard.
 
 from __future__ import annotations
 
-from jax.experimental.array_api._version import __array_api_version__ as __array_api_version__
-
 from jax.experimental.array_api import fft as fft
 from jax.experimental.array_api import linalg as linalg
 
 from jax.numpy import (
+    __array_api_version__ as __array_api_version__,
+    __array_namespace_info__ as __array_namespace_info__,
     abs as abs,
     acos as acos,
     acosh as acosh,
@@ -203,11 +203,3 @@ from jax.experimental.array_api._statistical_functions import (
     std as std,
     var as var,
 )
-
-from jax.experimental.array_api._utility_functions import (
-    __array_namespace_info__ as __array_namespace_info__,
-)
-
-from jax.experimental.array_api import _array_methods
-_array_methods.add_array_object_methods()
-del _array_methods